# Faster Pathname FileUtils methods