@schneems 2020-10-26
Thank you for the review Nobu. I understand akr maintains pathname but might not be on GitHub. Should I make a redmine issue for this proposal?