@nobu 2020-10-26
Feel free to file an issue.
The homepage in pathname.gemspec is https://github.com/ruby/pathname, but it's not found (maybe private repo?).