@schneems 2020-10-30
https://bugs.ruby-lang.org/issues/17287