@mauro-oto 2021-02-11
@akr is there any chance you could take a look at this PR in order to move it forward please? Thank you!! 

There also seems to be a few other `Pathname`-related PRs in the same state: https://github.com/ruby/ruby/pulls?q=pathname+is%3Aopen